pyproject.toml
[tool.poetry]
name = "my_moneyforward"
version = "0.1.0"
description = ""
authors = ["Hi-king <hikingko1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fire = "^0.4.0"
selenium = "^3.141.0"
webdriver-manager = "^3.4.2"
python-dateutil = "*"
pyyaml = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
